Remove suprasegmentals and syllabification
It was a good idea to support suprasegmentals and
syllabification. Unfortunately, the implementation here was very buggy
and practically unusable, to the extent that I avoided it even for a
set of heavily stress-dependent sound changes. Thus I am removing the
current implementation to make way for any better implementation which
I may think of in the future.