Refactor: move category expansion to outside parser
There are now two representations for sound changes: the immediate
result after parsing, and the result after all categories have been
inlined and expanded to lists of graphemes. This greatly simplifies
the code (since there’s now better separation of concerns), and should
allow it to become more flexible in the future.