Allow reporting intermediate results
As a side-effect, these changes mean that the input→output view now
contains glosses and spacing from the input. In fact this was also on
the TODO list, so it's a nice thing to have happened! But if this
turns out to be unwieldy it may yet be removed in the future.